third_party/blink/web_tests/wpt_internal/css/css-anchor-position/try-initial-transition.html

<!DOCTYPE html>
<title>CSS Anchor Positioning Test: Initial @try does not trigger a transition</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=cb>
  <div id=abs></div>
</div>
<style>
  #cb {
    position: relative;
    width: 100px;
    height: 100px;
    background: lightpink;
  }
  #abs {
    position: absolute;
    background: darkcyan;
    top: 10px;
    left: 10px;
    width: 25px;
    height: 25px;
    position-fallback: --pf;
    transition-property: top, left;
    transition-duration: 10s;
    transition-timing-function: steps(2, start);
  }
  @position-fallback --pf {
    @try {
      top: 50px;
      left: 50px;
    }
  }
</style>
<script>
  test(() => {
    assert_equals(getComputedStyle(abs).top, '50px');
    assert_equals(getComputedStyle(abs).left, '50px');
  }, 'No transition for initial style with @try');
</script>
